P3/backend/requirements.txt
asgiref==3.5.0 
djangorestframework==3.13.1 
djangorestframework-simplejwt==5.1.0 
Pillow==9.0.1 
PyJWT==2.3.0 
pytz==2021.3 
sqlparse==0.4.2
Django==4.0.3
tk==0.1.0